Add files via upload
## Summary
- split monolithic `command_handler.c` into modular source files
- add new header `command_actions.h` for command APIs
- update build script to compile new modules
